requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
newsapi-python==0.2.6
Pillow==8.0.1
pycparser==2.20
pygame==1.9.6
pykepler==1.0.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
virtualenv==20.1.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
